requirements.txt
Flask>=2.3.3
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0
instructor>=1.0
tiktoken>=0.7
PyYAML>=5.0
numpy>=2.2.6
torch>=2.7.0
torchvision>=0.22.0
torchaudio>=2.7.0
transformers>=4.52.3
kokoro>=0.9.4
soundfile>=0.12.1
langchain>=0.3.25
langchain-core>=0.3.62
langchain-community>=0.3.24
langchain-openai>=0.3.18
langchain-huggingface>=0.2.0
faiss-cpu>=1.11.0
sentence-transformers>=4.1.0
pytest>=8.3.5
pytest-snapshot>=0.9.0
psutil>=7.0.0
flake8>=7.0.0
vulture>=2.14
mypy>=1.16.0
pre-commit>=4.2.0
ruff>=0.5.0
